# openapi/fees/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    name: Kita Knirpsenstadt
  description: API für Beitragsverwaltung der Kita Knirpsenstadt
  title: Kita Knirpsenstadt Fees API
  version: 1.0.0
paths:
...
  /members:
    get:
      description: Get a paginated list of members with optional filtering and sorting
      parameters:
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
            default: 1
        - description: Items per page
          in: query
          name: perPage
          schema:
            type: integer
            default: 20
        - description: Search by name or member number
          in: query
          name: search
          schema:
            type: string
        - description: Sort by field
          in: query
          name: sortBy
          schema:
            type: string
            enum:
              - name
              - memberNumber
              - membershipStart
            default: name
        - description: Sort direction
          in: query
          name: sortDir
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
        - description: Filter by active status
          in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: List all members
      tags:
        - Members
    post:
      description: Create a new membership record
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
        description: Member data
        required: true
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create a new member
      tags:
        - Members
  "/members/{id}":
    delete:
      description: Delete a member by ID
      parameters:
        - description: Member ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member deleted successfully
        "400":
          description: Invalid member ID
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Delete a member
      tags:
        - Members
    get:
      description: Get detailed information about a specific member
      parameters:
        - description: Member ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Invalid member ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get a member by ID
      tags:
        - Members
    put:
      description: Update an existing member's information
      parameters:
        - description: Member ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
        description: Updated member data
        required: true
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Update a member
      tags:
        - Members
  /members/count:
    get:
      description: Number of members whose membership period covers the given date
        (asOf), including already inactive members. Defaults to today.
      parameters:
        - description: Reference date (YYYY-MM-DD), defaults to today
          in: query
          name: asOf
          schema:
            type: string
      responses:
        "200":
          description: Member count at reference date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberCountAsOf"
        "400":
          description: Invalid asOf date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Count members at a reference date
      tags:
        - Members
...
servers:
  - url: http://localhost:8081/api/fees/v1
  - url: https://localhost:8081/api/fees/v1
components:
...
  securitySchemes:
    BearerAuth:
      description: JWT Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    CreateMemberRequest:
      description: Request body for creating a new member
      properties:
        city:
          example: Berlin
          type: string
        email:
          example: hans.mueller@example.com
          type: string
        firstName:
          example: Hans
          type: string
        householdId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        lastName:
          example: Müller
          type: string
        memberNumber:
          example: M-2024-001
          type: string
        membershipEnd:
          example: 2024-12-31
          type: string
        membershipStart:
          example: 2024-01-01
          type: string
        phone:
          example: +49 123 456789
          type: string
        postalCode:
          example: "12345"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
      required:
        - firstName
        - lastName
        - membershipStart
      type: object
...
    ErrorResponse:
      description: Standard error response format
      properties:
        code:
          example: 400
          type: integer
        error:
          example: Bad Request
          type: string
        message:
          example: invalid request body
          type: string
      required:
        - code
        - error
        - message
      type: object
...
    Member:
      description: Member information
      properties:
        city:
          example: Berlin
          type: string
        email:
          example: hans.mueller@example.com
          type: string
        firstName:
          example: Hans
          type: string
        householdId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        isActive:
          example: true
          type: boolean
        lastName:
          example: Müller
          type: string
        memberNumber:
          example: M-2024-001
          type: string
        membershipEnd:
          example: 2024-12-31
          type: string
        membershipStart:
          example: 2024-01-01
          type: string
        phone:
          example: +49 123 456789
          type: string
        postalCode:
          example: "12345"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
      required:
        - firstName
        - id
        - isActive
        - lastName
        - membershipStart
      type: object
    MemberCountAsOf:
      description: Member count at a reference date, including inactive members
      properties:
        active:
          example: 40
          type: integer
        asOf:
          example: 2026-09-09
          type: string
        inactive:
          example: 5
          type: integer
        total:
          example: 45
          type: integer
      required:
        - active
        - asOf
        - inactive
        - total
      type: object
    MemberList:
      description: Paginated list of members
      properties:
        data:
          items:
            $ref: "#/components/schemas/Member"
          type: array
        page:
          example: 1
          type: integer
        perPage:
          example: 20
          type: integer
        total:
          example: 42
          type: integer
        totalPages:
          example: 3
          type: integer
      required:
        - data
        - page
        - perPage
        - total
        - totalPages
      type: object
...
    UpdateMemberRequest:
      description: Request body for updating a member
      properties:
        city:
          example: Berlin
          type: string
        email:
          example: hans.mueller@example.com
          type: string
        firstName:
          example: Hans
          type: string
        householdId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        isActive:
          example: true
          type: boolean
        lastName:
          example: Müller
          type: string
        membershipEnd:
          example: 2024-12-31
          type: string
        membershipStart:
          example: 2024-01-01
          type: string
        phone:
          example: +49 123 456789
          type: string
        postalCode:
          example: "12345"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
      type: object
...